animal_skin_disease_classifier/requirements.txt
Flask==3.0.3
tensorflow==2.15.0
opencv-python-headless==4.10.0.84
scikit-learn==1.5.2
numpy==1.26.4
Pillow==10.3.0
gunicorn==23.0.0
